agent: move gather test to an integration test
